feat(ngrx): remove app module
