--- src/app/app.module.ts
@@ -1,51 +1,0 @@
-import { HttpClient, HttpClientModule } from '@angular/common/http';
-import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
-import '@aotw/components';
-import { AotwIconRegistry } from '@aotw/components';
-import icons from '@aotw/core/dist/icons/icons.json';
-import { StoreModule } from '@ngrx/store';
-import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
-import { TranslateHttpLoader } from '@ngx-translate/http-loader';
-
-import { AppRoutingModule } from './app-routing.module';
-import { AppComponent } from './app.component';
-import { FooterComponent } from './components/footer';
-import { HeaderComponent } from './components/header';
-import { appReducer, layoutReducer } from './state/reducers';
-
-export function httpLoaderFactory(http: HttpClient): TranslateHttpLoader {
-  return new TranslateHttpLoader(http);
-}
-
-@NgModule({
-  declarations: [
-    AppComponent
-  ],
-  imports: [
-    BrowserModule,
-    FooterComponent,
-    HeaderComponent,
-    HttpClientModule,
-    StoreModule.forRoot({
-      app: appReducer,
-      layout: layoutReducer
-    }),
-    TranslateModule.forRoot({
-      defaultLanguage: 'en',
-      loader: {
-        provide: TranslateLoader,
-        useFactory: httpLoaderFactory,
-        deps: [HttpClient]
-      }
-    }),
-    AppRoutingModule
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
-export class AppModule {
-  public constructor() {
-    AotwIconRegistry.register(icons);
-  }
-}
