+++ src/main.ts
@@ -1,6 +1,41 @@
-import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
+import { HttpClient, provideHttpClient } from '@angular/common/http';
+import { importProvidersFrom } from '@angular/core';
+import { bootstrapApplication } from '@angular/platform-browser';
+import { provideRouter } from '@angular/router';
+import { provideStore } from '@ngrx/store';
+import { provideStoreDevtools } from '@ngrx/store-devtools';
+import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 
-import { AppModule } from './app/app.module';
+import { AppComponent } from './app/app.component';
+import { APP_ROUTES } from './app/app.routes';
+import { appReducer, layoutReducer } from './app/state/reducers';
 
-platformBrowserDynamic().bootstrapModule(AppModule)
+function httpLoaderFactory(http: HttpClient): TranslateHttpLoader {
+  return new TranslateHttpLoader(http);
+}
+
+bootstrapApplication(AppComponent, {
+  providers: [
+    importProvidersFrom(TranslateModule.forRoot({
+      defaultLanguage: 'en',
+      loader: {
+        provide: TranslateLoader,
+        useFactory: httpLoaderFactory,
+        deps: [HttpClient]
+      }
+    })),
+    provideHttpClient(),
+    provideRouter(APP_ROUTES),
+    provideStore({
+      app: appReducer,
+      layout: layoutReducer
+    }),
+    provideStoreDevtools({
+      maxAge: 25,
+      trace: false,
+      traceLimit: 75
+    })
+  ]
+})
   .catch(err => console.error(err));
